# lim/docs/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Lost In Motion
host: 0.0.0.0:8081
basePath: /api/v1
...
schemes:
  - http

#####################################################################
paths:
...
  /clients:
    patch:
      summary: Change password
      tags:
        - clients
      security:
        - JWTAuth: [ ]
      operationId: changePassword
      parameters:
        - in: body
          name: new_password
          description: "Новый пароль"
          required: true
          schema:
            $ref: '#/definitions/ClientChangePassword'
      responses:
        200:
          description: Success
        401:
          description: Authorization error
        403:
          description: No rights
        404:
          description: No such client
        500:
          description: Internal error
...
definitions:
...
  ClientChangePassword:
    type: object
    required:
      - password
    properties:
      password:
        type: string
...
securityDefinitions:
  JWTAuth:
    in:   header
    name: token
    type: apiKey 
...